Handle user input exceptions; print counter
Exception throwing and its immediate handling have been added
to the getUserInput function. Though not yet 100% precise, the user
is now notified of a problem with their input.

A minor bug has been fixed where clear() and ignore(...) was called
for cin instead of the stream we're working with, i.e. iss.

A counter has been added to the prompt message allowing user
to see what attempt they are on.